# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SAMT Unified Backend API
  version: 1.0.0
  description: Generated by scripts/generate-openapi.mjs from service OpenAPI docs exposed via API Gateway.
servers:
  - url: http://localhost:9080
...
paths:
...
  /api/sync/jira/issues:
    post:
      tags:
        - sync-controller
      summary: Trigger Jira issue sync for a single project config
      operationId: syncJiraIssues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sync_SyncRequestDto'
        required: true
      responses:
        '200':
          description: Jira sync completed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
  /api/sync/github/commits:
    post:
      tags:
        - sync-controller
      summary: Trigger GitHub commit sync for a single project config
      operationId: syncGithubCommits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sync_SyncRequestDto'
        required: true
      responses:
        '200':
          description: GitHub sync completed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
  /api/sync/all:
    post:
      tags:
        - sync-controller
      summary: Trigger Jira and GitHub sync in parallel for a single project config
      operationId: syncAll
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sync_SyncRequestDto'
        required: true
      responses:
        '200':
          description: All sync jobs completed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          description: At least one downstream sync failed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiProblemDetails'
      security:
        - bearerAuth: []
  /api/sync/jobs:
    get:
      tags:
        - sync-controller
      summary: List sync jobs
      operationId: listSyncJobs
      parameters:
        - name: projectConfigId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: jobType
          in: query
          required: false
          schema:
            type: string
            enum:
              - JIRA_ISSUES
              - JIRA_SPRINTS
              - GITHUB_COMMITS
              - GITHUB_PRS
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - RUNNING
              - COMPLETED
              - PARTIAL_FAILURE
              - FAILED
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
            maximum: 100000
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
            maximum: 100
            minimum: 1
        - name: x-schemathesis-unknown-property
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Sync jobs retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
  /api/sync/jobs/{syncJobId}:
    get:
      tags:
        - sync-controller
      summary: Get sync job by id
      operationId: getSyncJob
      parameters:
        - name: syncJobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Sync job retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/sync_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    sync_SyncRequestDto:
      type: object
      properties:
        projectConfigId:
          type: string
          format: uuid
      required:
        - projectConfigId
    sync_ApiResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    sync_ApiProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
...
    StandardError:
      type: object
      properties:
        type:
          type: string
          example: https://api.example.com/errors/validation-error
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: Invalid UUID
        instance:
          type: string
          example: /api/users/not-a-number
        timestamp:
          type: string
          format: date-time
...
  responses:
    Error400:
      description: 400 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error401:
      description: 401 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error403:
      description: 403 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error404:
      description: 404 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error405:
      description: 405 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error409:
      description: 409 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error415:
      description: 415 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error500:
      description: 500 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    identity_bearerAuth:
      type: http
      description: Enter JWT access token
      scheme: bearer
      bearerFormat: JWT
    userGroup_bearerAuth:
      type: http
      description: JWT token from identity-service
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT
    projectConfig_Bearer Authentication:
      type: http
      description: Enter your JWT token from Identity Service
      scheme: bearer
      bearerFormat: JWT
    sync_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
    report_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
...